[Fix] [FE] 삭제 버튼 확인 모달 추가
-

[Resolves] hotfix/delete-button
[Ref]
[Realted to]


Former-commit-id: cbade95b96c2ad0aa7a34636a455367629807638

+++ frontend/src/pages/board_page/mate_page/MateDetailPage.jsx
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { useParams, useNavigate, useLocation } from "react-router-dom";
+import Swal from "sweetalert2";
 import styled from "styled-components";
 import { useRecoilState } from "recoil";
 import axios from "../../../api/axios";
@@ -115,22 +116,38 @@ function MateDetailPage() {
   };
 
   const handleDeleteBtn = () => {
-    deleteMate(
-      boardId,
-      () => {
-        if (state && state.fromMypage) {
-          navigate("/mypage", {
-            state: {
-              user: state.user,
-              category: state.category,
-            },
-          });
-        } else {
-          navigate("/mate");
-        }
-      },
-      () => {},
-    );
+    Swal.fire({
+      html: `
+        <img src="https://raw.githubusercontent.com/Tarikul-Islam-Anik/Animated-Fluent-Emojis/master/Emojis/Animals/Bear.png" alt="Bear" width="100" height="100"/>
+        <p style='font-size:1.2rem; font-family:Pretendard-Bold;'>게시글을 삭제하시겠습니까?</p>
+      `,
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText:
+        "<div style='font-size:1rem; font-family:Pretendard-Regular;'>확인</div>",
+      cancelButtonText:
+        "<div style='font-size:1rem; font-family:Pretendard-Regular;'>취소</div>",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        deleteMate(
+          boardId,
+          () => {
+            if (state && state.fromMypage) {
+              navigate("/mypage", {
+                state: {
+                  user: state.user,
+                  category: state.category,
+                },
+              });
+            } else {
+              navigate("/mate");
+            }
+          },
+          () => {},
+        );
+      }
+    });
   };
 
   const renderEditDeleteButtons = () => {
